Fix reordering problem and resizing handles
Fixes a bug around resizing handles. The problem was that some state would
change (e.g. order of columns), and the resize handles wouldn't be recomputed.
Then if they were torn down, they would be in the wrong state and would cause
an error.

This also meant that sometimes after reordering columns, the wrong resize
handles would be applied.

The solution to this is (1) compute nextResizableColumn in a more sane way; and
(2) fix the resizing handle observer so that it fires at the right times.